15/11/2022 avances

+++ app/src/main/java/com/codekolih/producciontablet/clases/Produccion_Lista.java
@@ -44,7 +44,7 @@ public class Produccion_Lista implements Serializable  {
         return UserNameId;
     }
 
-    public float getProduccionId() {
+    public int getProduccionId() {
         return ProduccionId;
     }
